+++ src/Components/EmojiKeyBoard.jsx
@@ -1,122 +1,22 @@
+import { useEmojiKeyBoard } from "../hooks/useEmojiKeyBoard";
+
 function EmojiKeyBoard({ handleInputButtonClick }) {
-  const emojiCategories = {
-    smileys: ['😀', '😁', '😆', '🥹', '😅', '😂', '🤣', '🥲', '😊', '😇', '🙂', '🙃', '😉', '😌','😍', '🥰', '😘', '😋', '😛', '😝', '😜', '🤪', '🤨', '🧐', '🤓', '😎','🥸', '🤩', '🥳', '😏', '😒', '😞', '😔', '😟', '😕', '😣', '😖', '😫', '🥺', '😢', '😭', '😤', '😠', '😡', '🤬', '🤯', '😳', '🥵', '🥶', '😱', '😨', '😰', '🤗', '🤔', '🫣', '🤭', '🫢', '🤫', '🫠', '😶', '😐', '😑', '😬', '🙄', '😯', '😦', '🥱', '😴', '🤤', '😮‍💨', '🥴', '🤢', '🤮', '🤧', '😷', '🤒', '🤕', '🤑', '🤠', '😈', '👿', '👹', '👺', '🤡', '💩', '👻', '💀', '👽', '🤖', '🎃' ],
+    const emojiCategories = useEmojiKeyBoard();
 
-    animals: ['🐶', '🐱', '🐭', '🐹', '🐰', '🦊', '🐻', '🐼', '🐻‍❄️', '🐨', '🐯', '🦁', '🐮', '🐷', '🐽', '🐸', '🐵', '🙈', '🙉', '🙊', '🐒', '🐔', '🐧', '🐦', '🐤', '🐣', '🐥', '🦆', '🦅', '🦉', '🦇', '🐺', '🐗', '🐴', '🦄', '🐝', '🪱', '🐛', '🦋', '🐌', '🐞', '🐜', '🪰', '🪲', '🪳', '🦟', '🦗', '🕷', '🕸', '🦂', '🐢', '🐍', '🦎', '🦖', '🦕', '🐙', '🦑', '🦐', '🦞', '🦀', '🐡', '🐠', '🐟', '🐬', '🐳', '🐋', '🦈', '🐊', '🐅', '🐆', '🦓', '🦍', '🦧', '🦣', '🐘', '🦛', '🦏', '🐪', '🐫', '🦒', '🦘', '🦬', '🐃', '🐂', '🐄', '🐎', '🐖', '🐏', '🐑', '🦙', '🐐', '🦌', '🐕', '🐩', '🦮', '🐕‍🦺', '🐈', '🐓' ],
-
-    food: ['🍏', '🍎', '🍐', '🍊', '🍋', '🍌', '🍉', '🍇', '🍓', '🫐', '🍈', '🍒', '🍑', '🥭', '🍍', '🥥', '🥝', '🍅', '🫒', '🥑', '🍆', '🥔', '🥕', '🌽', '🌶', '🫑', '🥒', '🥬', '🥦', '🧄', '🧅', '🍄', '🥜', '🌰', '🍞', '🥐', '🥖', '🥨', '🥯', '🥞', '🧇', '🧀', '🍖', '🍗', '🥩', '🥓', '🍔', '🍟', '🍕', '🌭', '🥪', '🌮', '🌯', '🫔', '🥙', '🧆', '🥚', '🍳', '🥘', '🍲', '🫕', '🥣', '🥗', '🍿', '🧈', '🧂', '🥫', '🍱', '🍘', '🍙', '🍚', '🍛', '🍜', '🍝', '🍠', '🍢', '🍣', '🍤', '🍥', '🥮', '🍡', '🥟', '🥠', '🥡', '🦀', '🦞', '🦐', '🦑', '🦪', '🍦', '🍧', '🍨', '🍩', '🍪', '🎂', '🍰', '🧁', '🥧', '🍫', '🍬'],
-
-    drink: ['🍼', '🥛', '☕', '🫖', '🍵', '🧃', '🥤', '🍶', '🍺', '🍻', '🥂', '🍷', '🥃', '🍸', '🍹', '🧉', '🍾', '🧊', '🥄', '🍴', '🍽', '🥣', '🥡', '🥢'],
-
-    activities: ['⚽', '🏀', '🏈', '⚾', '🥎', '🎾', '🏐', '🏉', '🥏', '🎱', '🪀', '🏓', '🏸', '🏒', '🏑', '🥍', '🏏', '🪃', '🥅', '⛳', '🪁', '🏹', '🎣', '🤿', '🥊', '🥋', '🎽', '🛹', '🛷', '⛸', '🥌', '🎿', '⛷', '🏂', '🪂', '🏋️‍♀️', '🏋️', '🏋️‍♂️', '🤼‍♀️', '🤼', '🤼‍♂️', '🤸‍♀️', '🤸', '🤸‍♂️', '⛹️‍♀️', '⛹️', '⛹️‍♂️', '🤺', '🤾‍♀️', '🤾', '🤾‍♂️', '🏌️‍♀️', '🏌️', '🏌️‍♂️', '🏇', '🧘‍♀️', '🧘', '🧘‍♂️', '🏄‍♀️', '🏄', '🏄‍♂️', '🏊‍♀️', '🏊', '🏊‍♂️', '🤽‍♀️', '🤽', '🤽‍♂️', '🚣‍♀️', '🚣', '🚣‍♂️', '🧗‍♀️', '🧗', '🧗‍♂️', '🚵‍♀️', '🚵', '🚵'],
-
-    travel: [
-      '🚗','🚕','🚙','🚌','🚎','🏎','🚓',
-      '🚑','🚒','🚐','🚚','🚛','🚜','🦯',
-      '🦽','🦼','🛴','🚲','🛵','🏍','🛺',
-      '🚨','🚔','🚍','🚘','🚖','🚡','🚠',
-      '🚟','🚃','🚋','🚞','🚝','🚄','🚅',
-      '🚈','🚂','🚆','🚇','🚊','🚉','🛫',
-      '🛬','🛩','💺','🛰','🚀','🛸','🚁',
-      '🛶','⛵','🚤','🛥','🛳','🚢','⛽',
-      '🚧','🚦','🚥','🚏','🗺','🗿','🗽',
-      '🗼','🏰','🏯','🏟','🎡','🎢','🎠',
-      '⛲','🏖','🏝','🏜','🌋','🏔','🗻',
-      '🏕','⛺','🏠','🏡','🏘','🏚','🏗',
-    ],
-    objects: [
-      '⌚️','📱','📲','💻','🖥','🖨','🖱️',
-      '🖲️','🕹️','💿','📞','📷','📸','🎛️',
-      '⏱️','⏲️','⏰','🕰️','🕯️','💡','🔦',
-      '🏮','🪔','📔','📕','📖','📗','📘',
-      '📙','📚','📓','📒','📃','📜','📄',
-      '📰','🗞','📑','🔖','🏷','💰','💴',
-      '💵','💶','💷','💸','💳','🧾','💹',
-      '✉️','📧','📨','📩','📤','📥','📦',
-      '📫','📪','📬','📭','📮','📝','💼',
-      '📁','📂','🗂','📅','📆','🗒','🗓',
-      '📇','📈','📉','📊','📋','📌','📍',
-      '📎','🖇','📏','📐','✂️','🗃','🗄',
-      '🗑','🔒','🔓','🔏','🔐','🔑','🗝',
-      '🔨','🪓','⛏','⚒','🛠','🗡','⚔️',
-      '🔫','🏹','🛡','🪚','🔧','🪛','🔩',
-      '⚙️','🗜','⚖️','🦯','🔗','⛓','🪝',
-      '🧰','🧲','🪜','⚗️','🧪','🧫','🧬',
-      '🔬','🔭','📡','💉','🩸','💊','🩹',
-      '🩺','🚪','🛏','🛋','🪑','🚽','🚿',
-      '🛁','🪒','🧴','🧷','🧹','🧺','🧻',
-      '🪣','🧼','🪥','🧽','🧯','🛒','🚬',
-    ],
-    hearts: ['❤️', '🧡', '💛', '💚', '💙', '💜', '🖤', '🤍', '🤎', '💔', '❣️', '💕', '💞', '💓', '💗', '💖', '💘', '💝'],
-    symbols: [
-      '🔟','🔢','🔣','🔠','🔡','🔤','🅰️',
-      '🆎','🅱️','🆑','🆒','🆓','ℹ️','🆔',
-      'Ⓜ️','🆕','🆖','🅾️','🆗','🅿️','🆘',
-      '🆙','🆚','🈁','🈂️','🈷️','🈶','🈯',
-      '🉐','🈹','🈚','🈲','🉑','🈸','🈴',
-      '🈳','㊗️','㊙️','🈺','🈵','🔴','🟠',
-      '🟡','🟢','🔵','🟣','🟤','⚫','⚪',
-      '🟥','🟧','🟨','🟩','🟦','🟪','🟫',
-      '⬛','⬜','🔶','🔷','🔸','🔹','🔺','🔻',
-      '💠','🔘','🔳','🔲'
-    ],
-    flags: [
-      '🏁','🚩','🎌','🏴','🏳️','🏳️‍🌈','🏴‍☠️',
-      '🇦🇨','🇦🇩','🇦🇪','🇦🇫','🇦🇬','🇦🇮','🇦🇱',
-      '🇦🇲','🇦🇴','🇦🇶','🇦🇷','🇦🇸','🇦🇹','🇦🇺',
-      '🇦🇼','🇦🇽','🇦🇿','🇧🇦','🇧🇧','🇧🇩','🇧🇪',
-      '🇧🇫','🇧🇬','🇧🇭','🇧🇮','🇧🇯','🇧🇱','🇧🇲',
-      '🇧🇳','🇧🇴','🇧🇶','🇧🇷','🇧🇸','🇧🇹','🇧🇻',
-      '🇧🇼','🇧🇾','🇧🇿','🇨🇦','🇨🇨','🇨🇩','🇨🇫',
-      '🇨🇬','🇨🇭','🇨🇮','🇨🇰','🇨🇱','🇨🇲','🇨🇳',
-      '🇨🇴','🇨🇵','🇨🇷','🇨🇺','🇨🇻','🇨🇼','🇨🇽',
-      '🇨🇾','🇨🇿','🇩🇪','🇩🇬','🇩🇯','🇩🇰','🇩🇲',
-      '🇩🇴','🇩🇿','🇪🇦','🇪🇨','🇪🇪','🇪🇬','🇪🇭',
-      '🇪🇷','🇪🇸','🇪🇹','🇪🇺','🇫🇮','🇫🇯','🇫🇰',
-      '🇫🇲','🇫🇴','🇫🇷','🇬🇦','🇬🇧','🇬🇩','🇬🇪',
-      '🇬🇫','🇬🇬','🇬🇭','🇬🇮','🇬🇱','🇬🇲','🇬🇳',
-      '🇬🇵','🇬🇶','🇬🇷','🇬🇸','🇬🇹','🇬🇺','🇬🇼',
-      '🇬🇾','🇭🇰','🇭🇲','🇭🇳','🇭🇷','🇭🇹','🇭🇺',
-      '🇮🇨','🇮🇩','🇮🇪','🇮🇱','🇮🇲','🇮🇳','🇮🇴',
-      '🇮🇶','🇮🇷','🇮🇸','🇮🇹','🇯🇪','🇯🇲','🇯🇴',
-      '🇯🇵','🇰🇪','🇰🇬','🇰🇭','🇰🇮','🇰🇲','🇰🇳',
-      '🇰🇵','🇰🇷','🇰🇼','🇰🇾','🇰🇿','🇱🇦','🇱🇧',
-      '🇱🇨','🇱🇮','🇱🇰','🇱🇷','🇱🇸','🇱🇹','🇱🇺',
-      '🇱🇻','🇱🇾','🇲🇦','🇲🇨','🇲🇩','🇲🇪','🇲🇫',
-      '🇲🇬','🇲🇭','🇲🇰','🇲🇱','🇲🇲','🇲🇳','🇲🇴',
-      '🇲🇵','🇲🇶','🇲🇷','🇲🇸','🇲🇹','🇲🇺','🇲🇻',
-      '🇲🇼','🇲🇽','🇲🇾','🇲🇿','🇳🇦','🇳🇨','🇳🇪',
-      '🇳🇫','🇳🇬','🇳🇮','🇳🇱','🇳🇴','🇳🇵','🇳🇷',
-      '🇳🇺','🇳🇿','🇴🇲','🇵🇦','🇵🇪','🇵🇫','🇵🇬',
-      '🇵🇭','🇵🇰','🇵🇱','🇵🇲','🇵🇳','🇵🇷','🇵🇸',
-      '🇵🇹','🇵🇼','🇵🇾','🇶🇦','🇷🇪','🇷🇴','🇷🇸',
-      '🇷🇺','🇷🇼','🇸🇦','🇸🇧','🇸🇨','🇸🇩','🇸🇪',
-      '🇸🇬','🇸🇭','🇸🇮','🇸🇯','🇸🇰','🇸🇱','🇸🇲',
-      '🇸🇳','🇸🇴','🇸🇷','🇸🇸','🇸🇹','🇸🇻','🇸🇽',
-      '🇸🇾','🇸🇿','🇹🇦','🇹🇨','🇹🇩','🇹🇫','🇹🇬',
-      '🇹🇭','🇹🇯','🇹🇰','🇹🇱','🇹🇲','🇹🇳','🇹🇴',
-      '🇹🇷','🇹🇹','🇹🇻','🇹🇼','🇹🇿','🇺🇦','🇺🇬',
-      '🇺🇲','🇺🇳','🇺🇸','🇺🇾','🇺🇿','🇻🇦','🇻🇨',
-      '🇻🇪','🇻🇬','🇻🇮','🇻🇳','🇻🇺','🇼🇫','🇼🇸',
-      '🇽🇰','🇾🇪','🇾🇹','🇿🇦','🇿🇲','🇿🇼','🏴‍☠️',
-      '🏴󠁧󠁢󠁥󠁮󠁧󠁿','🏴󠁧󠁢󠁳󠁣󠁴󠁿','🏴󠁧󠁢󠁷󠁬󠁳󠁿',
-    ]
-  };
-
-  return (
-    <div className='emoji-keyboard'>
-      {Object.entries(emojiCategories).map(([category, emojis]) => (
-        <div key={category}>
-          <h3>{category}</h3>
-          {emojis.map((emoji, index) => (
-            <button key={index} onClick={() => handleInputButtonClick(emoji)}>
-              {emoji}
-            </button>
-          ))}
+    return (
+        <div className='emoji-keyboard'>
+            {Object.entries(emojiCategories).map(([category, emojis]) => (
+                <div key={category}>
+                    <h3>{category}</h3>
+                    {emojis.map((emoji, index) => (
+                        <button key={index} onClick={() => handleInputButtonClick(emoji)}>
+                            {emoji}
+                        </button>
+                    ))}
+                </div>
+            ))}
         </div>
-      ))}
-    </div>
-  );
+    );
 }
 
 export default EmojiKeyBoard;
