+++ src/hooks/useEmojiKeyBoard.js
@@ -1,0 +1,108 @@
+
+export const useEmojiKeyBoard = () => {
+    const emojiCategories = {
+        smileys: ['😀', '😁', '😆', '🥹', '😅', '😂', '🤣', '🥲', '😊', '😇', '🙂', '🙃', '😉', '😌','😍', '🥰', '😘', '😋', '😛', '😝', '😜', '🤪', '🤨', '🧐', '🤓', '😎','🥸', '🤩', '🥳', '😏', '😒', '😞', '😔', '😟', '😕', '😣', '😖', '😫', '🥺', '😢', '😭', '😤', '😠', '😡', '🤬', '🤯', '😳', '🥵', '🥶', '😱', '😨', '😰', '🤗', '🤔', '🫣', '🤭', '🫢', '🤫', '🫠', '😶', '😐', '😑', '😬', '🙄', '😯', '😦', '🥱', '😴', '🤤', '😮‍💨', '🥴', '🤢', '🤮', '🤧', '😷', '🤒', '🤕', '🤑', '🤠', '😈', '👿', '👹', '👺', '🤡', '💩', '👻', '💀', '👽', '🤖', '🎃' ],
+    
+        animals: ['🐶', '🐱', '🐭', '🐹', '🐰', '🦊', '🐻', '🐼', '🐻‍❄️', '🐨', '🐯', '🦁', '🐮', '🐷', '🐽', '🐸', '🐵', '🙈', '🙉', '🙊', '🐒', '🐔', '🐧', '🐦', '🐤', '🐣', '🐥', '🦆', '🦅', '🦉', '🦇', '🐺', '🐗', '🐴', '🦄', '🐝', '🪱', '🐛', '🦋', '🐌', '🐞', '🐜', '🪰', '🪲', '🪳', '🦟', '🦗', '🕷', '🕸', '🦂', '🐢', '🐍', '🦎', '🦖', '🦕', '🐙', '🦑', '🦐', '🦞', '🦀', '🐡', '🐠', '🐟', '🐬', '🐳', '🐋', '🦈', '🐊', '🐅', '🐆', '🦓', '🦍', '🦧', '🦣', '🐘', '🦛', '🦏', '🐪', '🐫', '🦒', '🦘', '🦬', '🐃', '🐂', '🐄', '🐎', '🐖', '🐏', '🐑', '🦙', '🐐', '🦌', '🐕', '🐩', '🦮', '🐕‍🦺', '🐈', '🐓' ],
+    
+        food: ['🍏', '🍎', '🍐', '🍊', '🍋', '🍌', '🍉', '🍇', '🍓', '🫐', '🍈', '🍒', '🍑', '🥭', '🍍', '🥥', '🥝', '🍅', '🫒', '🥑', '🍆', '🥔', '🥕', '🌽', '🌶', '🫑', '🥒', '🥬', '🥦', '🧄', '🧅', '🍄', '🥜', '🌰', '🍞', '🥐', '🥖', '🥨', '🥯', '🥞', '🧇', '🧀', '🍖', '🍗', '🥩', '🥓', '🍔', '🍟', '🍕', '🌭', '🥪', '🌮', '🌯', '🫔', '🥙', '🧆', '🥚', '🍳', '🥘', '🍲', '🫕', '🥣', '🥗', '🍿', '🧈', '🧂', '🥫', '🍱', '🍘', '🍙', '🍚', '🍛', '🍜', '🍝', '🍠', '🍢', '🍣', '🍤', '🍥', '🥮', '🍡', '🥟', '🥠', '🥡', '🦀', '🦞', '🦐', '🦑', '🦪', '🍦', '🍧', '🍨', '🍩', '🍪', '🎂', '🍰', '🧁', '🥧', '🍫', '🍬'],
+    
+        drink: ['🍼', '🥛', '☕', '🫖', '🍵', '🧃', '🥤', '🍶', '🍺', '🍻', '🥂', '🍷', '🥃', '🍸', '🍹', '🧉', '🍾', '🧊', '🥄', '🍴', '🍽', '🥣', '🥡', '🥢'],
+    
+        activities: ['⚽', '🏀', '🏈', '⚾', '🥎', '🎾', '🏐', '🏉', '🥏', '🎱', '🪀', '🏓', '🏸', '🏒', '🏑', '🥍', '🏏', '🪃', '🥅', '⛳', '🪁', '🏹', '🎣', '🤿', '🥊', '🥋', '🎽', '🛹', '🛷', '⛸', '🥌', '🎿', '⛷', '🏂', '🪂', '🏋️‍♀️', '🏋️', '🏋️‍♂️', '🤼‍♀️', '🤼', '🤼‍♂️', '🤸‍♀️', '🤸', '🤸‍♂️', '⛹️‍♀️', '⛹️', '⛹️‍♂️', '🤺', '🤾‍♀️', '🤾', '🤾‍♂️', '🏌️‍♀️', '🏌️', '🏌️‍♂️', '🏇', '🧘‍♀️', '🧘', '🧘‍♂️', '🏄‍♀️', '🏄', '🏄‍♂️', '🏊‍♀️', '🏊', '🏊‍♂️', '🤽‍♀️', '🤽', '🤽‍♂️', '🚣‍♀️', '🚣', '🚣‍♂️', '🧗‍♀️', '🧗', '🧗‍♂️', '🚵‍♀️', '🚵', '🚵'],
+    
+        travel: [
+          '🚗','🚕','🚙','🚌','🚎','🏎','🚓',
+          '🚑','🚒','🚐','🚚','🚛','🚜','🦯',
+          '🦽','🦼','🛴','🚲','🛵','🏍','🛺',
+          '🚨','🚔','🚍','🚘','🚖','🚡','🚠',
+          '🚟','🚃','🚋','🚞','🚝','🚄','🚅',
+          '🚈','🚂','🚆','🚇','🚊','🚉','🛫',
+          '🛬','🛩','💺','🛰','🚀','🛸','🚁',
+          '🛶','⛵','🚤','🛥','🛳','🚢','⛽',
+          '🚧','🚦','🚥','🚏','🗺','🗿','🗽',
+          '🗼','🏰','🏯','🏟','🎡','🎢','🎠',
+          '⛲','🏖','🏝','🏜','🌋','🏔','🗻',
+          '🏕','⛺','🏠','🏡','🏘','🏚','🏗',
+        ],
+        objects: [
+          '⌚️','📱','📲','💻','🖥','🖨','🖱️',
+          '🖲️','🕹️','💿','📞','📷','📸','🎛️',
+          '⏱️','⏲️','⏰','🕰️','🕯️','💡','🔦',
+          '🏮','🪔','📔','📕','📖','📗','📘',
+          '📙','📚','📓','📒','📃','📜','📄',
+          '📰','🗞','📑','🔖','🏷','💰','💴',
+          '💵','💶','💷','💸','💳','🧾','💹',
+          '✉️','📧','📨','📩','📤','📥','📦',
+          '📫','📪','📬','📭','📮','📝','💼',
+          '📁','📂','🗂','📅','📆','🗒','🗓',
+          '📇','📈','📉','📊','📋','📌','📍',
+          '📎','🖇','📏','📐','✂️','🗃','🗄',
+          '🗑','🔒','🔓','🔏','🔐','🔑','🗝',
+          '🔨','🪓','⛏','⚒','🛠','🗡','⚔️',
+          '🔫','🏹','🛡','🪚','🔧','🪛','🔩',
+          '⚙️','🗜','⚖️','🦯','🔗','⛓','🪝',
+          '🧰','🧲','🪜','⚗️','🧪','🧫','🧬',
+          '🔬','🔭','📡','💉','🩸','💊','🩹',
+          '🩺','🚪','🛏','🛋','🪑','🚽','🚿',
+          '🛁','🪒','🧴','🧷','🧹','🧺','🧻',
+          '🪣','🧼','🪥','🧽','🧯','🛒','🚬',
+        ],
+        hearts: ['❤️', '🧡', '💛', '💚', '💙', '💜', '🖤', '🤍', '🤎', '💔', '❣️', '💕', '💞', '💓', '💗', '💖', '💘', '💝'],
+        symbols: [
+          '🔟','🔢','🔣','🔠','🔡','🔤','🅰️',
+          '🆎','🅱️','🆑','🆒','🆓','ℹ️','🆔',
+          'Ⓜ️','🆕','🆖','🅾️','🆗','🅿️','🆘',
+          '🆙','🆚','🈁','🈂️','🈷️','🈶','🈯',
+          '🉐','🈹','🈚','🈲','🉑','🈸','🈴',
+          '🈳','㊗️','㊙️','🈺','🈵','🔴','🟠',
+          '🟡','🟢','🔵','🟣','🟤','⚫','⚪',
+          '🟥','🟧','🟨','🟩','🟦','🟪','🟫',
+          '⬛','⬜','🔶','🔷','🔸','🔹','🔺','🔻',
+          '💠','🔘','🔳','🔲'
+        ],
+        flags: [
+          '🏁','🚩','🎌','🏴','🏳️','🏳️‍🌈','🏴‍☠️',
+          '🇦🇨','🇦🇩','🇦🇪','🇦🇫','🇦🇬','🇦🇮','🇦🇱',
+          '🇦🇲','🇦🇴','🇦🇶','🇦🇷','🇦🇸','🇦🇹','🇦🇺',
+          '🇦🇼','🇦🇽','🇦🇿','🇧🇦','🇧🇧','🇧🇩','🇧🇪',
+          '🇧🇫','🇧🇬','🇧🇭','🇧🇮','🇧🇯','🇧🇱','🇧🇲',
+          '🇧🇳','🇧🇴','🇧🇶','🇧🇷','🇧🇸','🇧🇹','🇧🇻',
+          '🇧🇼','🇧🇾','🇧🇿','🇨🇦','🇨🇨','🇨🇩','🇨🇫',
+          '🇨🇬','🇨🇭','🇨🇮','🇨🇰','🇨🇱','🇨🇲','🇨🇳',
+          '🇨🇴','🇨🇵','🇨🇷','🇨🇺','🇨🇻','🇨🇼','🇨🇽',
+          '🇨🇾','🇨🇿','🇩🇪','🇩🇬','🇩🇯','🇩🇰','🇩🇲',
+          '🇩🇴','🇩🇿','🇪🇦','🇪🇨','🇪🇪','🇪🇬','🇪🇭',
+          '🇪🇷','🇪🇸','🇪🇹','🇪🇺','🇫🇮','🇫🇯','🇫🇰',
+          '🇫🇲','🇫🇴','🇫🇷','🇬🇦','🇬🇧','🇬🇩','🇬🇪',
+          '🇬🇫','🇬🇬','🇬🇭','🇬🇮','🇬🇱','🇬🇲','🇬🇳',
+          '🇬🇵','🇬🇶','🇬🇷','🇬🇸','🇬🇹','🇬🇺','🇬🇼',
+          '🇬🇾','🇭🇰','🇭🇲','🇭🇳','🇭🇷','🇭🇹','🇭🇺',
+          '🇮🇨','🇮🇩','🇮🇪','🇮🇱','🇮🇲','🇮🇳','🇮🇴',
+          '🇮🇶','🇮🇷','🇮🇸','🇮🇹','🇯🇪','🇯🇲','🇯🇴',
+          '🇯🇵','🇰🇪','🇰🇬','🇰🇭','🇰🇮','🇰🇲','🇰🇳',
+          '🇰🇵','🇰🇷','🇰🇼','🇰🇾','🇰🇿','🇱🇦','🇱🇧',
+          '🇱🇨','🇱🇮','🇱🇰','🇱🇷','🇱🇸','🇱🇹','🇱🇺',
+          '🇱🇻','🇱🇾','🇲🇦','🇲🇨','🇲🇩','🇲🇪','🇲🇫',
+          '🇲🇬','🇲🇭','🇲🇰','🇲🇱','🇲🇲','🇲🇳','🇲🇴',
+          '🇲🇵','🇲🇶','🇲🇷','🇲🇸','🇲🇹','🇲🇺','🇲🇻',
+          '🇲🇼','🇲🇽','🇲🇾','🇲🇿','🇳🇦','🇳🇨','🇳🇪',
+          '🇳🇫','🇳🇬','🇳🇮','🇳🇱','🇳🇴','🇳🇵','🇳🇷',
+          '🇳🇺','🇳🇿','🇴🇲','🇵🇦','🇵🇪','🇵🇫','🇵🇬',
+          '🇵🇭','🇵🇰','🇵🇱','🇵🇲','🇵🇳','🇵🇷','🇵🇸',
+          '🇵🇹','🇵🇼','🇵🇾','🇶🇦','🇷🇪','🇷🇴','🇷🇸',
+          '🇷🇺','🇷🇼','🇸🇦','🇸🇧','🇸🇨','🇸🇩','🇸🇪',
+          '🇸🇬','🇸🇭','🇸🇮','🇸🇯','🇸🇰','🇸🇱','🇸🇲',
+          '🇸🇳','🇸🇴','🇸🇷','🇸🇸','🇸🇹','🇸🇻','🇸🇽',
+          '🇸🇾','🇸🇿','🇹🇦','🇹🇨','🇹🇩','🇹🇫','🇹🇬',
+          '🇹🇭','🇹🇯','🇹🇰','🇹🇱','🇹🇲','🇹🇳','🇹🇴',
+          '🇹🇷','🇹🇹','🇹🇻','🇹🇼','🇹🇿','🇺🇦','🇺🇬',
+          '🇺🇲','🇺🇳','🇺🇸','🇺🇾','🇺🇿','🇻🇦','🇻🇨',
+          '🇻🇪','🇻🇬','🇻🇮','🇻🇳','🇻🇺','🇼🇫','🇼🇸',
+          '🇽🇰','🇾🇪','🇾🇹','🇿🇦','🇿🇲','🇿🇼','🏴‍☠️',
+          '🏴󠁧󠁢󠁥󠁮󠁧󠁿','🏴󠁧󠁢󠁳󠁣󠁴󠁿','🏴󠁧󠁢󠁷󠁬󠁳󠁿',
+        ]
+      };
+
+      return emojiCategories;
+}
